+++ docs/whatsnew.md
@@ -7,7 +7,7 @@
 - Added a new option for ORF finding, called `fast-naive` which is an internal very fast implementation.
 - Added the possibility of bypassing ORF finding altogether by providing prodigal outputs directly (or any other gene prediction in the right format)
 - Command line argument checking is more exhaustive instead of exiting at first error
-
+- Added `--quiet` flag to reduce the amount of output printed
 
 ## Version 1.4.0: long reads binning!